Extract - as_object, get_class_for to utils.py - base class to base.py
