+++ omise/api/resources/account.py
@@ -1,6 +1,6 @@
-from omise.api import *
-from omise.api.resources import *
-
+from .base import Base
+from .main_resource import MainResource
+from .utils import as_object
 
 class Account(MainResource, Base):
     """API class representing accounts details.
