+++ omise/api/resources/balance.py
@@ -1,5 +1,6 @@
-from omise.api import *
-from omise.api.resources import *
+from .base import Base
+from .main_resource import MainResource
+from .utils import as_object
 
 
 class Balance(MainResource, Base):
