+++ omise/api/resources/recipient.py
@@ -1,6 +1,7 @@
 import copy
-from omise.api import *
-from omise.api.resources import *
+from .base import Base
+from .main_resource import MainResource
+from .utils import as_object
 
 
 class Recipient(MainResource, Base):
