+++ omise/api/resources/refund.py
@@ -1,5 +1,6 @@
-from omise.api import *
-from omise.api.resources import *
+from .base import Base
+from .main_resource import MainResource
+from .lazy_collection import LazyCollection
 
 
 class Refund(MainResource, Base):
