+++ omise/api/resources/transaction.py
@@ -1,5 +1,7 @@
-from omise.api import *
-from omise.api.resources import *
+from .base import Base
+from .main_resource import MainResource
+from .utils import as_object
+from .lazy_collection import LazyCollection
 
 
 class Transaction(MainResource, Base):
